refactor(server): extract X/Twitter helpers → src/server/x.js
Pure move, zero behavior change. Pulls the three self-contained X
(Twitter) API primitives out of server.js into src/server/x.js:
  - buildXOAuthHeader — OAuth 1.0a HMAC-SHA1 Authorization header (8 refs)
  - uploadXMedia — image upload → media_id_string, v2/v1.1 branch (3 refs)
  - refreshXOAuth2Token — refresh_token → fresh access token (4 refs)

Why: continues the server.js dismemberment. These three are a cohesive
X-platform cluster with zero coupling to the rest of server.js — they
call no other local helpers; deps are only crypto (randomBytes,
createHmac) + Node globals (fetch, Buffer, process.env). buildGhostJWT
(a different platform) is left for its own cut.

What:
- New src/server/x.js exports all three (moved verbatim). Imports
  randomBytes/createHmac from crypto.
- server.js imports the three and keeps its own crypto import (those
  symbols are still used elsewhere). Inline defs → breadcrumb comments.

Test plan:
- node --check server.js -> OK
- npm run lint (no-undef gate) -> clean (confirms all three imported)
- route-inventory guard -> PASS (snapshot unchanged)
- vitest -> 60/60 pass (+3 new tests on buildXOAuthHeader: header shape,
  extraParams sign-but-don't-leak, signature is a 20-byte SHA-1 digest).
  uploadXMedia/refreshXOAuth2Token are network I/O — exercised via the
  live publish path, not unit-tested.

Rollback: revert; pure copy, restores inline defs exactly.
